requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements/requirements.in -o ./requirements/requirements.txt
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   instructor
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
appnope==0.1.4
    # via ipykernel
arxiv==2.1.3
    # via -r ./requirements/requirements.in
asttokens==2.4.1
    # via stack-data
attrs==24.2.0
    # via aiohttp
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.7
    # via langchain-community
debugpy==1.8.5
    # via ipykernel
decorator==5.1.1
    # via ipython
distro==1.9.0
    # via openai
docstring-parser==0.16
    # via instructor
executing==2.0.1
    # via stack-data
feedparser==6.0.11
    # via arxiv
fonttools==4.53.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
instructor==1.4.0
    # via -r ./requirements/requirements.in
ipykernel==6.29.5
    # via -r ./requirements/requirements.in
ipython==8.26.0
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.1.5
    # via -r ./requirements/requirements.in
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via -r ./requirements/requirements.in
jiter==0.4.2
    # via
    #   instructor
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jupyter-client==8.6.2
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
jupyterlab-widgets==3.0.13
    # via ipywidgets
kiwisolver==1.4.6
    # via matplotlib
langchain==0.2.14
    # via
    #   -r ./requirements/requirements.in
    #   langchain-community
langchain-community==0.2.12
    # via -r ./requirements/requirements.in
langchain-core==0.2.33
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.2.2
    # via langchain
langsmith==0.1.99
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.21.3
    # via dataclasses-json
matplotlib==3.9.2
    # via -r ./requirements/requirements.in
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via ipykernel
numpy==1.26.4
    # via
    #   contourpy
    #   langchain
    #   langchain-community
    #   matplotlib
openai==1.41.0
    # via
    #   -r ./requirements/requirements.in
    #   instructor
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   ipykernel
    #   langchain-core
    #   marshmallow
    #   matplotlib
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==10.4.0
    # via
    #   -r ./requirements/requirements.in
    #   matplotlib
platformdirs==4.2.2
    # via jupyter-core
prompt-toolkit==3.0.47
    # via ipython
psutil==6.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pydantic==2.8.2
    # via
    #   -r ./requirements/requirements.in
    #   instructor
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via
    #   instructor
    #   pydantic
pygments==2.18.0
    # via
    #   ipython
    #   rich
pymupdf==1.24.10
    # via -r ./requirements/requirements.in
pymupdfb==1.24.10
    # via pymupdf
pyparsing==3.1.4
    # via matplotlib
pypdf==4.3.1
    # via -r ./requirements/requirements.in
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
pyzmq==26.1.1
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.3
    # via
    #   arxiv
    #   langchain
    #   langchain-community
    #   langsmith
rich==13.8.0
    # via
    #   instructor
    #   typer
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==1.4.27
    # via
    #   langchain
    #   langchain-community
stack-data==0.6.3
    # via ipython
tenacity==8.5.0
    # via
    #   instructor
    #   langchain
    #   langchain-community
    #   langchain-core
tornado==6.4.1
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.66.5
    # via openai
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typer==0.12.5
    # via instructor
typing-extensions==4.12.2
    # via
    #   ipython
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==1.26.20
    # via requests
uv==0.2.37
    # via -r ./requirements/requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
widgetsnbextension==4.0.13
    # via ipywidgets
yarl==1.9.4
    # via aiohttp
